Improve metadata handling for RIT
Some fields are often skipped in some simulations' metadata
Some fields are present but their values are left empty

We now insert numerical values in both of these scenarios,
the value we currently insert is zero (0). For some of the
fields, this value is the most reasonable one (e.g. for x,y
components of spin for non-spinning or aligned-spin configs,
and orbital eccentricity),
while for all others its a clearly absurd choice which would
indicate its artifical insertion